move BinaryLabelCounter and BinaryConfusionMatrixImpl to evaluation.binary
